Move ValueProfile from costs.py to values.py
